Code Refactoring: Refactor editor.js, remove code comments, add new comments, simplify code
